refactor(testkit): dedupe StubParamStore for core tests (#1202)
