+++ icn/crates/icn-core/src/supervisor/decision_executor.rs
@@ -584,105 +584,6 @@ mod tests {
     fn make_test_executor() -> (Arc<DecisionExecutor>, Arc<MemoryExecutionStore>) {
         use icn_kernel_api::protocol_params::*;
 
-        // Minimal param store stub
-        struct StubParamStore;
-        impl ProtocolParameterStore for StubParamStore {
-            fn get(&self, _: &str) -> Result<Option<ProtocolParameter>> {
-                Ok(None)
-            }
-            fn get_effective(
-                &self,
-                _: &str,
-                _: Option<&str>,
-                _: Option<&str>,
-            ) -> Result<Option<ProtocolParameter>> {
-                Ok(None)
-            }
-            fn set(
-                &self,
-                _: ProtocolParameter,
-                _: Option<String>,
-                _: Option<String>,
-            ) -> Result<()> {
-                Ok(())
-            }
-            fn list(&self) -> Result<Vec<ProtocolParameter>> {
-                Ok(vec![])
-            }
-            fn list_by_category(&self, _: &str) -> Result<Vec<ProtocolParameter>> {
-                Ok(vec![])
-            }
-            fn get_history(&self, _: &str) -> Result<Vec<ParameterChange>> {
-                Ok(vec![])
-            }
-            fn get_history_paginated(
-                &self,
-                _: &str,
-                _: usize,
-                _: usize,
-            ) -> Result<(Vec<ParameterChange>, usize)> {
-                Ok((vec![], 0))
-            }
-            fn prune_history(&self, _: &str, _: usize) -> Result<usize> {
-                Ok(0)
-            }
-            fn delete(&self, _: &str) -> Result<()> {
-                Ok(())
-            }
-            fn exists(&self, _: &str) -> Result<bool> {
-                Ok(false)
-            }
-            fn count(&self) -> Result<usize> {
-                Ok(0)
-            }
-            fn total_history_count(&self) -> Result<usize> {
-                Ok(0)
-            }
-            fn validate(
-                &self,
-                _: &str,
-                _: &ParameterValue,
-            ) -> std::result::Result<(), ParameterValidationError> {
-                Ok(())
-            }
-            fn list_scoped_parameters(&self) -> Result<Vec<ProtocolParameter>> {
-                Ok(vec![])
-            }
-            fn delete_scoped_parameter(&self, _: &str, _: &ParameterScope) -> Result<bool> {
-                Ok(false)
-            }
-            fn add_pending_change(&self, _: PendingParameterChange) -> Result<()> {
-                Ok(())
-            }
-            fn get_pending_change(
-                &self,
-                _: &PendingChangeId,
-            ) -> Result<Option<PendingParameterChange>> {
-                Ok(None)
-            }
-            fn list_pending_changes(&self) -> Result<Vec<PendingParameterChange>> {
-                Ok(vec![])
-            }
-            fn list_pending_changes_for_parameter(
-                &self,
-                _: &str,
-            ) -> Result<Vec<PendingParameterChange>> {
-                Ok(vec![])
-            }
-            fn get_changes_due_before(&self, _: u64) -> Result<Vec<PendingParameterChange>> {
-                Ok(vec![])
-            }
-            fn update_pending_change(&self, _: PendingParameterChange) -> Result<()> {
-                Ok(())
-            }
-            fn cancel_pending_change(&self, _: &PendingChangeId, _: &str) -> Result<()> {
-                Ok(())
-            }
-            fn count_pending_changes(&self) -> Result<usize> {
-                Ok(0)
-            }
-        }
-
         let param_store = Arc::new(StubParamStore);
         let kernel_executor =
             Arc::new(super::super::governance_executor::KernelGovernanceExecutor::new(param_store));
